Add motives and improve test for BA model

+++ Complex Network/CommonLibrary/Model/Result/AnalizeResult.cs
@@ -24,7 +24,7 @@ namespace RandomGraph.Common.Model.Result
             DistancesBetweenEigenValues = new SortedDictionary<double, int>();
             Cycles = new SortedDictionary<int, long>();
             //CyclesCount = new SortedDictionary<int, int>();
-            MotivesCount = new SortedDictionary<int, int>();
+            MotivesCount = new SortedDictionary<int, float>();
             //CyclesCountForBA = new SortedDictionary<int, long>();
         }
 
@@ -48,7 +48,7 @@ namespace RandomGraph.Common.Model.Result
 
         //public SortedDictionary<int, int> CyclesCount { get; set; }
 
-        public SortedDictionary<int, int> MotivesCount { get; set; }
+        public SortedDictionary<int, float> MotivesCount { get; set; }
 
         public SortedDictionary<int, long> Cycles { get; set; }
